refactor: extract the code of task operation message, and add remove task message
